requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver, can be replaced with other DB driver

# Security and Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# GraphQL
strawberry-graphql[fastapi]==0.211.1

# File Handling
aiofiles==23.2.1
Pillow==10.1.0  # For image processing

# Utilities
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1